+++ device.js
@@ -19,8 +19,8 @@ module.exports = [
     'name': 'Chrome',
     'userAgent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/66.0.3359.139 Safari/537.36',
     'viewport': {
-      'width': 1920,
-      'height': 1080,
+      'screenWidth': 1920,
+      'screenHeight': 1080,
       'mobile': false
     },
     'isTouch': true
@@ -29,8 +29,8 @@ module.exports = [
     'name': 'iPad Pro',
     'userAgent': 'Mozilla/5.0 (iPad; CPU OS 11_0 like Mac OS X) AppleWebKit/604.1.34 (KHTML, like Gecko) Version/11.0 Mobile/15A5341f Safari/604.1',
     'viewport': {
-      'width': 1024,
-      'height': 1366,
+      'screenWidth': 1024,
+      'screenHeight': 1366,
       'deviceScaleFactor': 1,
       'mobile': true
     },
@@ -40,8 +40,8 @@ module.exports = [
     'name': 'iPad landscape',
     'userAgent': 'Mozilla/5.0 (iPad; CPU OS 11_0 like Mac OS X) AppleWebKit/604.1.34 (KHTML, like Gecko) Version/11.0 Mobile/15A5341f Safari/604.1',
     'viewport': {
-      'width': 1024,
-      'height': 768,
+      'screenWidth': 1024,
+      'screenHeight': 768,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -51,8 +51,8 @@ module.exports = [
     'name': 'iPad Mini',
     'userAgent': 'Mozilla/5.0 (iPad; CPU OS 11_0 like Mac OS X) AppleWebKit/604.1.34 (KHTML, like Gecko) Version/11.0 Mobile/15A5341f Safari/604.1',
     'viewport': {
-      'width': 768,
-      'height': 1024,
+      'screenWidth': 768,
+      'screenHeight': 1024,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -62,8 +62,8 @@ module.exports = [
     'name': 'iPad Mini landscape',
     'userAgent': 'Mozilla/5.0 (iPad; CPU OS 11_0 like Mac OS X) AppleWebKit/604.1.34 (KHTML, like Gecko) Version/11.0 Mobile/15A5341f Safari/604.1',
     'viewport': {
-      'width': 1024,
-      'height': 768,
+      'screenWidth': 1024,
+      'screenHeight': 768,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -73,8 +73,8 @@ module.exports = [
     'name': 'iPad Pro landscape',
     'userAgent': 'Mozilla/5.0 (iPad; CPU OS 11_0 like Mac OS X) AppleWebKit/604.1.34 (KHTML, like Gecko) Version/11.0 Mobile/15A5341f Safari/604.1',
     'viewport': {
-      'width': 1366,
-      'height': 1024,
+      'screenWidth': 1366,
+      'screenHeight': 1024,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -84,8 +84,8 @@ module.exports = [
     'name': 'iPhone 4',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 7_1_2 like Mac OS X) AppleWebKit/537.51.2 (KHTML, like Gecko) Version/7.0 Mobile/11D257 Safari/9537.53',
     'viewport': {
-      'width': 320,
-      'height': 480,
+      'screenWidth': 320,
+      'screenHeight': 480,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -95,8 +95,8 @@ module.exports = [
     'name': 'iPhone 4 landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 7_1_2 like Mac OS X) AppleWebKit/537.51.2 (KHTML, like Gecko) Version/7.0 Mobile/11D257 Safari/9537.53',
     'viewport': {
-      'width': 480,
-      'height': 320,
+      'screenWidth': 480,
+      'screenHeight': 320,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -106,8 +106,8 @@ module.exports = [
     'name': 'iPhone 5',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_1 like Mac OS X) AppleWebKit/603.1.30 (KHTML, like Gecko) Version/10.0 Mobile/14E304 Safari/602.1',
     'viewport': {
-      'width': 320,
-      'height': 568,
+      'screenWidth': 320,
+      'screenHeight': 568,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -117,8 +117,8 @@ module.exports = [
     'name': 'iPhone 5 landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_1 like Mac OS X) AppleWebKit/603.1.30 (KHTML, like Gecko) Version/10.0 Mobile/14E304 Safari/602.1',
     'viewport': {
-      'width': 568,
-      'height': 320,
+      'screenWidth': 568,
+      'screenHeight': 320,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -128,8 +128,8 @@ module.exports = [
     'name': 'iPhone 6',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 375,
-      'height': 667,
+      'screenWidth': 375,
+      'screenHeight': 667,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -139,8 +139,8 @@ module.exports = [
     'name': 'iPhone 6 landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 667,
-      'height': 375,
+      'screenWidth': 667,
+      'screenHeight': 375,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -150,8 +150,8 @@ module.exports = [
     'name': 'iPhone 6 Plus',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 414,
-      'height': 736,
+      'screenWidth': 414,
+      'screenHeight': 736,
       'deviceScaleFactor': 3,
       'mobile': true
     },
@@ -161,8 +161,8 @@ module.exports = [
     'name': 'iPhone 6 Plus landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 736,
-      'height': 414,
+      'screenWidth': 736,
+      'screenHeight': 414,
       'deviceScaleFactor': 3,
       'mobile': true,
     },
@@ -172,8 +172,8 @@ module.exports = [
     'name': 'iPhone 7',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 375,
-      'height': 667,
+      'screenWidth': 375,
+      'screenHeight': 667,
       'deviceScaleFactor': 2,
       'mobile': true,
     },
@@ -183,8 +183,8 @@ module.exports = [
     'name': 'iPhone 7 landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 667,
-      'height': 375,
+      'screenWidth': 667,
+      'screenHeight': 375,
       'deviceScaleFactor': 2,
       'mobile': true,
     },
@@ -194,8 +194,8 @@ module.exports = [
     'name': 'iPhone 7 Plus',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 414,
-      'height': 736,
+      'screenWidth': 414,
+      'screenHeight': 736,
       'deviceScaleFactor': 3,
       'mobile': true,
     },
@@ -205,8 +205,8 @@ module.exports = [
     'name': 'iPhone 7 Plus landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 736,
-      'height': 414,
+      'screenWidth': 736,
+      'screenHeight': 414,
       'deviceScaleFactor': 3,
       'mobile': true,
       'isTouch': true,
@@ -216,8 +216,8 @@ module.exports = [
     'name': 'iPhone 8',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 375,
-      'height': 667,
+      'screenWidth': 375,
+      'screenHeight': 667,
       'deviceScaleFactor': 2,
       'mobile': true,
       'isTouch': true,
@@ -227,8 +227,8 @@ module.exports = [
     'name': 'iPhone 8 landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 667,
-      'height': 375,
+      'screenWidth': 667,
+      'screenHeight': 375,
       'deviceScaleFactor': 2,
       'mobile': true,
       'isTouch': true,
@@ -238,8 +238,8 @@ module.exports = [
     'name': 'iPhone 8 Plus',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 414,
-      'height': 736,
+      'screenWidth': 414,
+      'screenHeight': 736,
       'deviceScaleFactor': 3,
       'mobile': true
     },
@@ -249,8 +249,8 @@ module.exports = [
     'name': 'iPhone 8 Plus landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 736,
-      'height': 414,
+      'screenWidth': 736,
+      'screenHeight': 414,
       'deviceScaleFactor': 3,
       'mobile': true
     },
@@ -260,8 +260,8 @@ module.exports = [
     'name': 'iPhone SE',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_1 like Mac OS X) AppleWebKit/603.1.30 (KHTML, like Gecko) Version/10.0 Mobile/14E304 Safari/602.1',
     'viewport': {
-      'width': 320,
-      'height': 568,
+      'screenWidth': 320,
+      'screenHeight': 568,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -271,8 +271,8 @@ module.exports = [
     'name': 'iPhone SE landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_3_1 like Mac OS X) AppleWebKit/603.1.30 (KHTML, like Gecko) Version/10.0 Mobile/14E304 Safari/602.1',
     'viewport': {
-      'width': 568,
-      'height': 320,
+      'screenWidth': 568,
+      'screenHeight': 320,
       'deviceScaleFactor': 2,
       'mobile': true
     },
@@ -282,8 +282,8 @@ module.exports = [
     'name': 'iPhone X',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 375,
-      'height': 812,
+      'screenWidth': 375,
+      'screenHeight': 812,
       'deviceScaleFactor': 3,
       'mobile': true
     },
@@ -293,119 +293,9 @@ module.exports = [
     'name': 'iPhone X landscape',
     'userAgent': 'Mozilla/5.0 (iPhone; CPU iPhone OS 11_0 like Mac OS X) AppleWebKit/604.1.38 (KHTML, like Gecko) Version/11.0 Mobile/15A372 Safari/604.1',
     'viewport': {
-      'width': 812,
-      'height': 375,
+      'screenWidth': 812,
+      'screenHeight': 375,
       'deviceScaleFactor': 3,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 10',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 6.0.1; Nexus 10 Build/MOB31T) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Safari/537.36',
-    'viewport': {
-      'width': 800,
-      'height': 1280,
-      'deviceScaleFactor': 2,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 10 landscape',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 6.0.1; Nexus 10 Build/MOB31T) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Safari/537.36',
-    'viewport': {
-      'width': 1280,
-      'height': 800,
-      'deviceScaleFactor': 2,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 4',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 4.4.2; Nexus 4 Build/KOT49H) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 384,
-      'height': 640,
-      'deviceScaleFactor': 2,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 4 landscape',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 4.4.2; Nexus 4 Build/KOT49H) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 640,
-      'height': 384,
-      'deviceScaleFactor': 2,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 5',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 6.0; Nexus 5 Build/MRA58N) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 360,
-      'height': 640,
-      'deviceScaleFactor': 3,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 5 landscape',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 6.0; Nexus 5 Build/MRA58N) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 640,
-      'height': 360,
-      'deviceScaleFactor': 3,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 5X',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 8.0.0; Nexus 5X Build/OPR4.170623.006) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 412,
-      'height': 732,
-      'deviceScaleFactor': 2.625,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 5X landscape',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 8.0.0; Nexus 5X Build/OPR4.170623.006) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 732,
-      'height': 412,
-      'deviceScaleFactor': 2.625,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 6',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 7.1.1; Nexus 6 Build/N6F26U) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 412,
-      'height': 732,
-      'deviceScaleFactor': 3.5,
-      'mobile': true
-    },
-    'isTouch': true,
-  },
-  {
-    'name': 'Nexus 6 landscape',
-    'userAgent': 'Mozilla/5.0 (Linux; Android 7.1.1; Nexus 6 Build/N6F26U) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
-    'viewport': {
-      'width': 732,
-      'height': 412,
-      'deviceScaleFactor': 3.5,
       'mobile': true
     },
     'isTouch': true,
@@ -414,8 +304,8 @@ module.exports = [
     'name': 'Nexus 6P',
     'userAgent': 'Mozilla/5.0 (Linux; Android 8.0.0; Nexus 6P Build/OPP3.170518.006) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3452.0 Mobile Safari/537.36',
     'viewport': {
-      'width': 412,
-      'height': 732,
+      'screenWidth': 412,
+      'screenHeight': 732,
       'deviceScaleFactor': 3.5,
       'mobile': true
     },