+++ History.md
@@ -28,3 +28,9 @@
 ### 0.9.2
 
 * 优化双工消息队列，将无上下级依赖的同级消息由异步发送改为同步发送
+
+### 0.10.0
+
+* 增加viewport.screenWidth、viewport.screenHeight参数，用于定义屏幕分辨率
+
+* viewport.width、viewport.height在缺省状态下会用screenWidth、screenHeight进行填充